# api/openapi.json
{
 "openapi": "3.0.3",
 "info": {
 "title": "xrelease API",
 "description": "Self-hosted release notifier — management API and inbound webhooks.\n\nStarted with `xrelease serve`. Polling continues in the background; webhooks provide low-latency push for GitHub, GitLab, Gitea, Codeberg, Bitbucket, and Docker Hub.\n\nConfiguration is split: `bootstrap.toml` (infra) + `app/releases.yaml` (sources/teams/notifiers). Source kinds (`type` in app config): `github`, `codeberg`, `gitea`, `gitlab`, `bitbucket`, `docker`, `ghcr`, `quay`, `ecr`, `feed`, `pypi`, `npm`, `yarn`, `cargo`, `maven`, `nuget`, `hex`, `rubygems`, `packagist`, `cpan`, `artifacthub`.",
 "version": "0.1.0",
 "license": {
 "name": "Apache-2.0",
 "url": "https://www.apache.org/licenses/LICENSE-2.0"
 }
 },
 "servers": [
 {
 "url": "http://127.0.0.1:8080",
 "description": "Default local bind (`[api].listen`)"
 }
 ],
...
 "paths": {
...
 "/api/v1/auth/login": {
 "post": {
 "tags": [
 "auth"
 ],
 "summary": "Local username/password login",
 "operationId": "postAuthLogin",
 "requestBody": {
 "required": true,
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/AuthLoginRequest"
 }
 }
 }
 },
 "responses": {
 "200": {
 "description": "Session JWT",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/AuthLoginResponse"
 }
 }
 }
 },
 "401": {
 "$ref": "#/components/responses/Unauthorized"
 }
 }
 }
 },
 "/api/v1/auth/oidc/sync": {
 "post": {
 "tags": [
 "auth"
 ],
 "summary": "Upsert OIDC user and assign role from claims",
 "operationId": "postAuthOidcSync",
 "security": [
 {
 "bearerAuth": []
 }
 ],
 "responses": {
 "200": {
 "description": "Persisted user profile",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/AuthUserView"
 }
 }
 }
 },
 "401": {
 "$ref": "#/components/responses/Unauthorized"
 }
 }
 }
 },
 "/api/v1/auth/me": {
 "get": {
 "tags": [
 "auth"
 ],
 "summary": "Current authenticated principal",
 "operationId": "getAuthMe",
 "security": [
 {
 "bearerAuth": []
 }
 ],
 "responses": {
 "200": {
 "description": "Principal + optional user",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/AuthMeResponse"
 }
 }
 }
 },
 "401": {
 "$ref": "#/components/responses/Unauthorized"
 }
 }
 }
 },
 "/api/v1/auth/logout": {
 "post": {
 "tags": [
 "auth"
 ],
 "summary": "Revoke the caller's live sessions (bumps session_version)",
 "operationId": "logout",
 "security": [
 {
 "bearerAuth": []
 }
 ],
 "responses": {
 "200": {
 "description": "revoked (true for a local session, false otherwise)",
 "content": {
 "application/json": {
 "schema": {
 "type": "object",
 "properties": {
 "revoked": {
 "type": "boolean"
 }
 }
 }
 }
 }
 },
 "401": {
 "description": "Missing/invalid or revoked bearer",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/ErrorResponse"
 }
 }
 }
 }
 }
 }
 },
 "/api/v1/auth/users": {
 "get": {
 "tags": [
 "auth"
 ],
 "summary": "List application users (local + OIDC)",
 "description": "Admin-only directory of persisted identities. OIDC rows appear after the first successful `/auth/oidc/sync`.",
 "operationId": "listAuthUsers",
 "security": [
 {
 "bearerAuth": []
 }
 ],
 "responses": {
 "200": {
 "description": "User directory",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/AuthUserListResponse"
 }
 }
 }
 },
 "401": {
 "$ref": "#/components/responses/Unauthorized"
 },
 "403": {
 "$ref": "#/components/responses/Forbidden"
 }
 }
 },
 "post": {
 "tags": [
 "auth"
 ],
 "summary": "Create a local username/password user",
 "description": "Admin-only. Requires local auth to be enabled. OIDC users appear via `/auth/oidc/sync` (auto-links matching local email) or admin `POST …/users/{id}/oidc`.",
 "operationId": "createAuthUser",
 "security": [
 {
 "bearerAuth": []
 }
 ],
 "requestBody": {
 "required": true,
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/AuthCreateUserRequest"
 }
 }
 }
 },
 "responses": {
 "200": {
 "description": "Created user",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/AuthUserView"
 }
 }
 }
 },
 "400": {
 "$ref": "#/components/responses/BadRequest"
 },
 "401": {
 "$ref": "#/components/responses/Unauthorized"
 },
 "403": {
 "$ref": "#/components/responses/Forbidden"
 },
 "409": {
 "$ref": "#/components/responses/Conflict"
 }
 }
 }
 },
 "/api/v1/auth/users/{id}/oidc": {
 "post": {
 "tags": [
 "auth"
 ],
 "summary": "Link or unlink an OIDC subject on a local user",
 "description": "Admin-only. Attaches `oidc_sub` to an existing local (password) user so the same row can sign in via SSO. `auth_source` stays `local`. Pass a blank/null `oidc_sub` to unlink. Auto-link also happens on `/auth/oidc/sync` when emails match.",
 "operationId": "linkAuthUserOidc",
 "security": [
 {
 "bearerAuth": []
 }
 ],
 "parameters": [
 {
 "name": "id",
 "in": "path",
 "required": true,
 "schema": {
 "type": "integer",
 "format": "int64"
 }
 }
 ],
 "requestBody": {
 "required": true,
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/AuthLinkOidcRequest"
 }
 }
 }
 },
 "responses": {
 "200": {
 "description": "Updated user",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/AuthUserView"
 }
 }
 }
 },
 "400": {
 "$ref": "#/components/responses/BadRequest"
 },
 "401": {
 "$ref": "#/components/responses/Unauthorized"
 },
 "403": {
 "$ref": "#/components/responses/Forbidden"
 },
 "404": {
 "$ref": "#/components/responses/NotFound"
 },
 "409": {
 "$ref": "#/components/responses/Conflict"
 }
 }
 }
 },
...
 },
 "components": {
 "securitySchemes": {
 "bearerAuth": {
 "type": "http",
 "scheme": "bearer",
 "description": "Static API key (`[api].api_key` / `XRELEASE_API_KEY`) or OIDC access token JWT (`[api.oidc]` / `XRELEASE_OIDC_ISSUER`). Required when either auth method is configured."
 }
 },
 "responses": {
 "Unauthorized": {
 "description": "Authentication failed",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/ErrorResponse"
 }
 }
 }
 },
 "BadRequest": {
 "description": "Invalid payload",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/ErrorResponse"
 }
 }
 }
 },
 "Forbidden": {
 "description": "Authenticated but insufficient role",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/ErrorResponse"
 }
 }
 }
 },
 "Conflict": {
 "description": "Request conflicts with current state",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/ErrorResponse"
 }
 }
 }
 },
 "NotFound": {
 "description": "No matching configured source",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/ErrorResponse"
 }
 }
 }
 },
...
 },
 "schemas": {
...
 "ErrorResponse": {
 "type": "object",
 "properties": {
 "error": {
 "type": "string"
 }
 },
 "required": [
 "error"
 ]
 },
...
 "AppRole": {
 "type": "string",
 "enum": [
 "admin",
 "operator",
 "viewer"
 ]
 },
 "AuthLoginRequest": {
 "type": "object",
 "required": [
 "username",
 "password"
 ],
 "properties": {
 "username": {
 "type": "string"
 },
 "password": {
 "type": "string",
 "format": "password"
 }
 }
 },
 "AuthUserView": {
 "type": "object",
 "required": [
 "id",
 "role",
 "auth_source",
 "created_at"
 ],
 "properties": {
 "id": {
 "type": "integer",
 "format": "int64"
 },
 "username": {
 "type": "string",
 "nullable": true
 },
 "email": {
 "type": "string",
 "nullable": true
 },
 "display_name": {
 "type": "string",
 "nullable": true
 },
 "role": {
 "$ref": "#/components/schemas/AppRole"
 },
 "auth_source": {
 "type": "string",
 "description": "`local` (password) or `oidc` (created/synced from IdP).",
 "enum": [
 "local",
 "oidc"
 ]
 },
 "oidc_sub": {
 "type": "string",
 "nullable": true,
 "description": "OIDC subject (`sub`) when the user is linked to an IdP identity. Omitted when unset."
 },
 "created_at": {
 "type": "string",
 "format": "date-time"
 },
 "last_login_at": {
 "type": "string",
 "format": "date-time",
 "nullable": true,
 "description": "Last successful login. Omitted when the user has never signed in."
 },
 "organization_roles": {
 "type": "object",
 "description": "Per-organization role grants from OIDC scoped aliases (e.g. xrelease-admin:platform). Omitted when empty.",
 "additionalProperties": {
 "$ref": "#/components/schemas/AppRole"
 }
 }
 }
 },
 "AuthUserListResponse": {
 "type": "object",
 "required": [
 "users"
 ],
 "properties": {
 "users": {
 "type": "array",
 "items": {
 "$ref": "#/components/schemas/AuthUserView"
 }
 }
 }
 },
 "AuthCreateUserRequest": {
 "type": "object",
 "required": [
 "username",
 "password",
 "role"
 ],
 "properties": {
 "username": {
 "type": "string",
 "minLength": 1,
 "maxLength": 64
 },
 "password": {
 "type": "string",
 "format": "password",
 "minLength": 8
 },
 "role": {
 "$ref": "#/components/schemas/AppRole"
 },
 "email": {
 "type": "string",
 "nullable": true
 },
 "display_name": {
 "type": "string",
 "nullable": true
 }
 }
 },
 "AuthLinkOidcRequest": {
 "type": "object",
 "properties": {
 "oidc_sub": {
 "type": "string",
 "nullable": true,
 "description": "IdP subject to attach. Null or blank unlinks SSO from the local user."
 }
 }
 },
 "AuthLoginResponse": {
 "type": "object",
 "required": [
 "access_token",
 "token_type",
 "expires_in",
 "role",
 "user"
 ],
 "properties": {
 "access_token": {
 "type": "string"
 },
 "token_type": {
 "type": "string"
 },
 "expires_in": {
 "type": "integer",
 "format": "int64"
 },
 "role": {
 "$ref": "#/components/schemas/AppRole"
 },
 "user": {
 "$ref": "#/components/schemas/AuthUserView"
 }
 }
 },
...
 "AuthMeResponse": {
 "type": "object",
 "required": [
 "method",
 "organization_roles"
 ],
 "properties": {
 "method": {
 "type": "string"
 },
 "role": {
 "allOf": [
 {
 "$ref": "#/components/schemas/AppRole"
 }
 ],
 "nullable": true,
 "description": "Instance-wide role. Api-key principals are always admin."
 },
 "organization_roles": {
 "type": "object",
 "description": "Per-organization grants from OIDC scoped aliases (admin:platform). Empty for local and api-key.",
 "additionalProperties": {
 "$ref": "#/components/schemas/AppRole"
 }
 },
 "user": {
 "allOf": [
 {
 "$ref": "#/components/schemas/AuthUserView"
 }
 ],
 "nullable": true
 }
 }
 },
...
 }
 }
}